+++ src/prefixes/AdminPrefix.js
@@ -1,6 +1,7 @@
 const Products = require("./admin/Products");
 const Roles = require("./admin/Roles");
 const Users = require("./admin/Users");
+const { query, nullToUndefined } = require("@simpleview/sv-graphql-client");
 
 const apis = {
 	products : Products,
@@ -12,13 +13,43 @@ class AdminPrefix {
 	constructor({ graphUrl, graphServer }) {
 		this.name = "admin";
 		
+		this._graphUrl = graphUrl;
+		this._graphServer = graphServer;
+		
 		for(let [name, Api] of Object.entries(apis)) {
 			const api = new Api({ graphUrl, name, graphServer : graphServer });
 			this[name] = api.find.bind(api);
 			this[`${name}_upsert`] = api.upsert.bind(api);
 			this[`${name}_remove`] = api.remove.bind(api);
-			this[`${name}_import`] = api.import.bind(api);
 		}
+	}
+	async users_import({fields, context }) {
+		context = context || this._graphServer.context;
+		
+		const variables = {
+			acct_id : context.acct_id
+		}
+		
+		const response = await query({
+			query : `
+				mutation($acct_id: String!) {
+					admin(acct_id: $acct_id) {
+						users_import{
+							${fields}
+						}
+					}
+				}
+			`,
+			variables,
+			url : this._graphUrl,
+			token : context.token
+		});
+		
+		const returnData = response.admin.users_import;
+		
+		nullToUndefined(returnData);
+		
+		return returnData;
 	}
 }
